+++ src/main/java/com/iexec/core/task/update/TaskUpdateManager.java
@@ -116,7 +116,6 @@ class TaskUpdateManager {
         );
     }
 
-    @SuppressWarnings("DuplicateBranchesInSwitch")
     void updateTask(String chainTaskId) {
         Optional<Task> optional = taskService.getTaskByChainTaskId(chainTaskId);
         if (optional.isEmpty()) {
@@ -128,11 +127,10 @@ class TaskUpdateManager {
         boolean isFinalDeadlinePossible =
                 !TaskStatus.getStatusesWhereFinalDeadlineIsImpossible().contains(currentStatus);
         if (isFinalDeadlinePossible && new Date().after(task.getFinalDeadline())) {
-            updateTaskStatusAndSave(task, FINAL_DEADLINE_REACHED);
             // Eventually should fire a "final deadline reached" notification to worker,
             // but here let's just trigger an toFailed(task) leading to a failed status
             // which will itself fire a generic "abort" notification
-            toFailed(task);
+            toFailed(task, FINAL_DEADLINE_REACHED);
             return;
         }
 
@@ -147,24 +145,15 @@ class TaskUpdateManager {
                 initialized2Running(task);
                 initializedOrRunning2ContributionTimeout(task);
                 break;
-            case INITIALIZE_FAILED:
-                toFailed(task);
-                break;
             case RUNNING:
                 running2Finalized2Completed(task); // running2Finalized2Completed must be the first call to prevent other transition execution
                 running2ConsensusReached(task);
                 running2RunningFailed(task);
                 initializedOrRunning2ContributionTimeout(task);
                 break;
-            case RUNNING_FAILED:
-                toFailed(task);
-                break;
             case CONSENSUS_REACHED:
                 consensusReached2AtLeastOneReveal2ResultUploading(task);
                 consensusReached2Reopening(task);
-                break;
-            case CONTRIBUTION_TIMEOUT:
-                toFailed(task);
                 break;
             case AT_LEAST_ONE_REVEALED:
                 requestUpload(task);
@@ -175,9 +164,6 @@ class TaskUpdateManager {
             case REOPENED:
                 updateTaskStatusAndSave(task, INITIALIZED);
                 break;
-            case REOPEN_FAILED:
-                toFailed(task);
-                break;
             case RESULT_UPLOADING:
                 resultUploading2Uploaded(task);
                 resultUploading2UploadTimeout(task);
@@ -185,18 +171,18 @@ class TaskUpdateManager {
             case RESULT_UPLOADED:
                 resultUploaded2Finalizing(task);
                 break;
-            case RESULT_UPLOAD_TIMEOUT:
-                toFailed(task);
-                break;
             case FINALIZING:
                 finalizing2Finalized2Completed(task);
                 break;
             case FINALIZED:
                 finalizedToCompleted(task);
                 break;
+            case INITIALIZE_FAILED:
+            case RUNNING_FAILED:
+            case CONTRIBUTION_TIMEOUT:
+            case REOPEN_FAILED:
+            case RESULT_UPLOAD_TIMEOUT:
             case FINALIZE_FAILED:
-                toFailed(task);
-                break;
             case FINAL_DEADLINE_REACHED:
                 toFailed(task);
                 break;
@@ -206,25 +192,49 @@ class TaskUpdateManager {
         }
     }
 
-    Task updateTaskStatusAndSave(Task task, TaskStatus newStatus) {
-        return updateTaskStatusAndSave(task, newStatus, null);
+    /**
+     * Creates one or several task status changes for the task before committing all of them to the database.
+     *
+     * @param task     The task
+     * @param statuses List of statuses to append to the task {@code dateStatusList}
+     */
+    void updateTaskStatusesAndSave(Task task, TaskStatus... statuses) {
+        TaskStatus initialStatus = task.getCurrentStatus();
+        TaskStatus lastStatus = statuses[statuses.length - 1];
+        for (TaskStatus newStatus : statuses) {
+            log.info("Create TaskStatusChange succeeded [chainTaskId:{}, currentStatus:{}, newStatus:{}]",
+                    task.getChainTaskId(), task.getCurrentStatus(), newStatus);
+            task.changeStatus(newStatus, null);
+        }
+        saveTask(task, initialStatus, lastStatus);
     }
 
-    Task updateTaskStatusAndSave(Task task, TaskStatus newStatus, ChainReceipt chainReceipt) {
-        TaskStatus currentStatus = task.getCurrentStatus();
-        task.changeStatus(newStatus, chainReceipt);
-        Optional<Task> savedTask = taskService.updateTask(task);
+    void updateTaskStatusAndSave(Task task, TaskStatus newStatus) {
+        updateTaskStatusAndSave(task, newStatus, null);
+    }
 
+    void updateTaskStatusAndSave(Task task, TaskStatus newStatus, ChainReceipt chainReceipt) {
+        TaskStatus initialStatus = task.getCurrentStatus();
+        task.changeStatus(newStatus, chainReceipt);
+        saveTask(task, initialStatus, newStatus);
+    }
+
+    /**
+     * Saves the task to the database.
+     *
+     * @param task          The task
+     * @param currentStatus The current status in database
+     * @param newStatus     The new status in database after save
+     */
+    void saveTask(Task task, TaskStatus currentStatus, TaskStatus newStatus) {
+        Optional<Task> savedTask = taskService.updateTask(task);
         // `savedTask.isPresent()` should always be true if the task exists in the repository.
         if (savedTask.isPresent()) {
             updateMetricsAfterStatusUpdate(currentStatus, newStatus);
             log.info("UpdateTaskStatus succeeded [chainTaskId:{}, currentStatus:{}, newStatus:{}]", task.getChainTaskId(), currentStatus, newStatus);
-            return savedTask.get();
         } else {
-            log.warn("UpdateTaskStatus failed. Chain Task is probably unknown." +
-                            " [chainTaskId:{}, currentStatus:{}, wishedStatus:{}]",
+            log.warn("UpdateTaskStatus failed. Chain Task is probably unknown [chainTaskId:{}, currentStatus:{}, wishedStatus:{}]",
                     task.getChainTaskId(), currentStatus, newStatus);
-            return null;
         }
     }
 
@@ -253,8 +263,7 @@ class TaskUpdateManager {
             Optional<String> smsUrl = smsService.getVerifiedSmsUrl(task.getChainTaskId(), task.getTag());
             if (smsUrl.isEmpty()) {
                 log.error("Couldn't get verified SMS url [chainTaskId: {}]", task.getChainTaskId());
-                updateTaskStatusAndSave(task, INITIALIZE_FAILED);
-                updateTaskStatusAndSave(task, FAILED);
+                toFailed(task, INITIALIZE_FAILED);
                 return;
             }
             task.setSmsUrl(smsUrl.get()); //SMS URL source of truth for the task
@@ -271,8 +280,7 @@ class TaskUpdateManager {
                     if (enclaveChallenge.isEmpty()) {
                         log.error("Can't initialize task, enclave challenge is empty [chainTaskId:{}]",
                                 chainTaskId);
-                        updateTaskStatusAndSave(task, INITIALIZE_FAILED);
-                        updateTask(chainTaskId);
+                        toFailed(task, INITIALIZE_FAILED);
                         return;
                     }
                     task.setEnclaveChallenge(enclaveChallenge.get());
@@ -282,8 +290,7 @@ class TaskUpdateManager {
                 }, () -> {
                     log.error("Failed to request initialize on blockchain [chainTaskId:{}]",
                             task.getChainTaskId());
-                    updateTaskStatusAndSave(task, INITIALIZE_FAILED);
-                    updateTaskStatusAndSave(task, FAILED);
+                    toFailed(task, INITIALIZE_FAILED);
                 });
     }
 
@@ -306,8 +313,7 @@ class TaskUpdateManager {
                     }
                     log.error("Initialization failed on blockchain (tx reverted) [chainTaskId:{}]",
                             task.getChainTaskId());
-                    updateTaskStatusAndSave(task, INITIALIZE_FAILED);
-                    updateTaskStatusAndSave(task, FAILED);
+                    toFailed(task, INITIALIZE_FAILED);
                 }, () -> log.error("Unable to check initialization on blockchain " +
                                 "(likely too long), should use a detector [chainTaskId:{}]",
                         task.getChainTaskId()));
@@ -407,12 +413,11 @@ class TaskUpdateManager {
         } else if (nbReplicatesWithContributeAndFinalizeStatus > 1) {
             log.error("Too many replicates in ContributeAndFinalize status"
                     + " [chainTaskId:{}, nbReplicates:{}]", chainTaskId, nbReplicatesWithContributeAndFinalizeStatus);
-            toFailed(task);
+            toFailed(task, RUNNING_FAILED);
             return;
         }
 
-        updateTaskStatusAndSave(task, FINALIZED);
-        finalizedToCompleted(task);
+        toFinalizedToCompleted(task);
     }
 
     void initializedOrRunning2ContributionTimeout(Task task) {
@@ -421,8 +426,7 @@ class TaskUpdateManager {
         boolean isNowAfterContributionDeadline = task.getContributionDeadline() != null && new Date().after(task.getContributionDeadline());
 
         if (isInitializedOrRunningTask && isNowAfterContributionDeadline) {
-            updateTaskStatusAndSave(task, CONTRIBUTION_TIMEOUT);
-            updateTaskStatusAndSave(task, FAILED);
+            updateTaskStatusesAndSave(task, CONTRIBUTION_TIMEOUT, FAILED);
             applicationEventPublisher.publishEvent(new ContributionTimeoutEvent(task.getChainTaskId()));
         }
     }
@@ -473,8 +477,7 @@ class TaskUpdateManager {
         // If all alive workers have failed on this task, its computation should be stopped.
         // It could denote that the task is wrong
         // - e.g. failing script, dataset can't be retrieved, app can't be downloaded, ...
-        updateTaskStatusAndSave(task, RUNNING_FAILED);
-        updateTaskStatusAndSave(task, FAILED);
+        updateTaskStatusesAndSave(task, RUNNING_FAILED, FAILED);
         applicationEventPublisher.publishEvent(new TaskRunningFailedEvent(task.getChainTaskId()));
     }
 
@@ -595,9 +598,8 @@ class TaskUpdateManager {
                 && new Date().after(task.getFinalDeadline());
 
         if (isTaskInResultUploading && isNowAfterFinalDeadline) {
-            updateTaskStatusAndSave(task, RESULT_UPLOAD_TIMEOUT);
             applicationEventPublisher.publishEvent(new ResultUploadTimeoutEvent(task.getChainTaskId()));
-            updateTaskStatusAndSave(task, FAILED);
+            toFailed(task, RESULT_UPLOAD_TIMEOUT);
         }
     }
 
@@ -660,8 +662,7 @@ class TaskUpdateManager {
                 }, () -> {
                     log.error("Failed to request finalize on blockchain [chainTaskId:{}]",
                             task.getChainTaskId());
-                    updateTaskStatusAndSave(task, FINALIZE_FAILED);
-                    updateTaskStatusAndSave(task, FAILED);
+                    toFailed(task, FINALIZE_FAILED);
                 });
     }
 
@@ -672,14 +673,12 @@ class TaskUpdateManager {
                     if (Boolean.TRUE.equals(isSuccess)) {
                         log.info("Finalized on blockchain (tx mined) [chainTaskId:{}]",
                                 task.getChainTaskId());
-                        updateTaskStatusAndSave(task, FINALIZED, null);
-                        finalizedToCompleted(task);
+                        toFinalizedToCompleted(task);
                         return;
                     }
                     log.error("Finalization failed on blockchain (tx reverted) [chainTaskId:{}]",
                             task.getChainTaskId());
-                    updateTaskStatusAndSave(task, FINALIZE_FAILED);
-                    updateTaskStatusAndSave(task, FAILED);
+                    toFailed(task, FINALIZE_FAILED);
                 }, () -> log.error("Unable to check finalization on blockchain " +
                                 "(likely too long), should use a detector [chainTaskId:{}]",
                         task.getChainTaskId()));
@@ -693,8 +692,18 @@ class TaskUpdateManager {
         applicationEventPublisher.publishEvent(new TaskCompletedEvent(task));
     }
 
+    void toFinalizedToCompleted(Task task) {
+        updateTaskStatusesAndSave(task, FINALIZED, COMPLETED);
+        applicationEventPublisher.publishEvent(new TaskCompletedEvent(task));
+    }
+
     void toFailed(Task task) {
         updateTaskStatusAndSave(task, FAILED);
+        applicationEventPublisher.publishEvent(new TaskFailedEvent(task.getChainTaskId()));
+    }
+
+    void toFailed(Task task, TaskStatus reason) {
+        updateTaskStatusesAndSave(task, reason, FAILED);
         applicationEventPublisher.publishEvent(new TaskFailedEvent(task.getChainTaskId()));
     }